learned how to use cmsis and set up RCC

+++ startup.c
@@ -56,7 +56,7 @@ void reset_handler(void){
 
     }
     uint32_t bss_size = (uint32_t)&_ebss - (uint32_t)&_sbss;
-    uint8_t *bss = (uint8_t*)&_sbss;
+    uint32_t *bss = (uint32_t*)&_sbss;
 
     for(uint32_t i = 0; i < bss_size; i++){
         bss[i] = 0;
